chore: split cpu initialization as per OS
