+++ metric/cpu/cpu.go
@@ -24,7 +24,6 @@ import (
 	"github.com/elastic/elastic-agent-libs/mapstr"
 	"github.com/elastic/elastic-agent-libs/opt"
 	"github.com/elastic/elastic-agent-system-metrics/metric"
-	"github.com/elastic/elastic-agent-system-metrics/metric/system/resolve"
 )
 
 // CPU manages the CPU metrics from /proc/stat
@@ -77,23 +76,6 @@ func (cpu CPU) Total() uint64 {
 	// it's generally safe to blindly sum these up,
 	// As we're just trying to get a total of all CPU time.
 	return opt.SumOptUint(cpu.User, cpu.Nice, cpu.Sys, cpu.Idle, cpu.Wait, cpu.Irq, cpu.SoftIrq, cpu.Stolen)
-}
-
-/*
-The below code implements a "metrics tracker" that gives us the ability to
-calculate CPU percentages, as we average usage across a time period.
-*/
-
-// Monitor is used to monitor the overall CPU usage of the system over time.
-type Monitor struct {
-	lastSample CPUMetrics
-	Hostfs     resolve.Resolver
-}
-
-// New returns a new CPU metrics monitor
-// Hostfs is only relevant on linux and freebsd.
-func New(hostfs resolve.Resolver) *Monitor {
-	return &Monitor{Hostfs: hostfs}
 }
 
 // Fetch collects a new sample of the CPU usage metrics.
